requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
arch==7.2.0
astropy==7.0.1
astropy-iers-data==0.2025.4.21.0.37.6
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clarabel==0.10.0
click==8.1.7
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
curl_cffi==0.10.0
cvxpy==1.6.5
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fqdn==1.5.1
frozendict==2.4.6
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.44
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-leaflet==0.19.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multitasking==0.0.11
narwhals==1.36.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.2
orjson==3.10.7
osqp==1.0.3
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.17.9
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyerfa==2.0.1.5
Pygments==2.18.0
pymdown-extensions==10.11.2
pyparsing==3.1.4
pyperclip==1.9.0
pyshp==2.3.1
python-box==7.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ratelim==0.1.6
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
Riskfolio-Lib==7.0.0
rpds-py==0.20.0
rsa==4.9
ruff==0.6.9
scikit-learn==1.6.1
scipy==1.15.2
scooby==0.10.0
scs==3.2.7.post2
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
statsmodels==0.14.4
streamlit==1.44.1
streamlit-option-menu==0.4.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.3.0
toml==0.10.2
tomlkit==0.12.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.13
XlsxWriter==3.2.3
xyzservices==2024.9.0
yfinance==0.2.56
